+++ composer.json
@@ -15,7 +15,8 @@
         "symplify/easy-testing": "^10.0",
         "symplify/easy-coding-standard": "^11.2",
         "rector/rector": "^0.15.19",
-        "laravel/framework": "^10.2"
+        "laravel/framework": "^10.2",
+        "symplify/easy-ci": "11.2.0.72"
     },
     "autoload": {
         "psr-4": {
