+++ src/Compiler/PhpContentExtractor.php
@@ -4,21 +4,6 @@ declare(strict_types=1);
 
 namespace TomasVotruba\Bladestan\Compiler;
 
-use function array_map;
-use function array_unshift;
-use function explode;
-use function implode;
-use function preg_match;
-use function preg_match_all;
-use function rtrim;
-use function str_replace;
-use function str_starts_with;
-use function strip_tags;
-use function token_name;
-use function trim;
-
-use const PHP_EOL;
-
 final class PhpContentExtractor
 {
     /**
