+++ src/TemplateCompiler/ErrorSkipper.php
@@ -7,10 +7,19 @@ namespace TomasVotruba\Bladestan\TemplateCompiler;
 use PHPStan\Analyser\Error;
 
 /**
- * @see \Bladestan\TemplateCompiler\Tests\ErrorSkipperTest
+ * @see \TomasVotruba\Bladestan\Tests\TemplateCompiler\ErrorSkipperTest
  */
 final class ErrorSkipper
 {
+    /**
+     * @var string[]
+     */
+    private const ERRORS_TO_IGNORE_REGEXES = [
+        '#Call to function unset\(\) contains undefined variable \$loop#',
+        '#Variable \$loop in PHPDoc tag @var does not exist#',
+        '#Anonymous function has an unused use (.*?)#',
+    ];
+
     /**
      * @param Error[] $errors
      * @param string[] $errorIgnores
@@ -22,7 +31,7 @@ final class ErrorSkipper
 
         foreach ($errors as $error) {
             foreach ($errorIgnores as $errorIgnore) {
-                $result = preg_match($errorIgnore, $error->getMessage(), $matches);
+                $result = preg_match($errorIgnore, $error->getMessage());
                 if ($result !== false) {
                     continue 2;
                 }
@@ -32,5 +41,22 @@ final class ErrorSkipper
         }
 
         return $filteredErrors;
+    }
+
+    /**
+     * @param Error[] $ruleErrors
+     * @return Error[]
+     */
+    public function filterErrors(array $ruleErrors): array
+    {
+        foreach ($ruleErrors as $key => $ruleError) {
+            foreach (self::ERRORS_TO_IGNORE_REGEXES as $errorToIgnoreRegex) {
+                if (! preg_match($errorToIgnoreRegex, $ruleError->getMessage())) {
+                    continue;
+                }
+
+                unset($ruleErrors[$key]);
+            }
+        }
     }
 }
